requirements.txt
streamlit
python-dotenv
pandas
ta
requests
python-binance
pycoingecko